크기 조정 및 환경 변경

+++ Unity/RobotArm_FK_Example/Assets/Script/View/SliderValue.cs
@@ -54,7 +54,7 @@ public class SliderValue : MonoBehaviour {
 
     public void OnChangedKinematics()
     {
-        switch (slider.value)
+        switch ((int)slider.value)
         {
             case 0:
                 titleText.text = "Forward Kinematics";
